fix: view certificate preview for teak
Replace edxapp wrapper with implementation for course module to fix Certificate preview for Teak.

fccn/nau-technical#661
