+++ nau_openedx_extensions/coursecertificate/test_settings.py
@@ -48,7 +48,7 @@ DATABASES = {
 }
 
 NAU_COURSE_MODULE = (
-    "nau_openedx_extensions.edxapp_wrapper.backends.course_module_l_v1_tests"
+    "nau_openedx_extensions.edxapp_wrapper.backends.course_module"
 )
 NAU_EMAIL_MODULE = (
     "nau_openedx_extensions.edxapp_wrapper.backends.email_module_l_v1_tests"
