--- nau_openedx_extensions/edxapp_wrapper/backends/course_module_l_v1_tests.py
@@ -1,8 +1,0 @@
-""" Course block backend abstraction for tests"""
-
-from unittest.mock import Mock
-
-
-def get_other_course_settings(course_id):   # pylint: disable=unused-argument
-    """Get Other Course Settings Mock."""
-    return Mock()
